requirements.txt
python-telegram-bot==20.7
gTTS==2.5.1